requirements.txt
pyqrcode
pillow
pypng
